# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rigoblock",
    "x-logo": {
      "url": "https://trader.rigoblock.com/favicon.ico",
      "altText": "Rigoblock"
    },
    "description": "AI-powered DeFi trading API for autonomous agents. Protected by the x402 v2 payment protocol — pay per request in USDC on Base mainnet, no API key required.\n\n## x402 v2 Payment Flow\n1. Make a request → server responds `402 Payment Required` and may include a `PAYMENT-REQUIRED` header describing accepted payment requirements\n2. Create the payment payload for USDC on Base (chain 8453), typically with `@x402/core`\n3. Retry with the `X-Payment` header → receive structured data; successful 2xx responses may include a `PAYMENT-RESPONSE` settlement receipt\n\n**Use the [@x402/core](https://www.npmjs.com/package/@x402/core) SDK** — it handles payment flow automatically.\n\n**Payment:** USDC on Base (eip155:8453) — CDP facilitator at api.cdp.coinbase.com\n\n**Integration guide & agent-to-agent architecture:** https://github.com/RigoBlock/agentic-operator/blob/main/AGENTS.md\n\n**Supported chains:** Ethereum (1), Base (8453), Arbitrum (42161), Optimism (10), Polygon (137), BNB (56), Unichain (130)",
    "version": "1.0.0",
    "contact": {
      "name": "Rigoblock",
      "url": "https://rigoblock.com",
      "email": "gab@rigoblock.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
...
    }
  ],
  "paths": {
...
    "/api/settings/slippage": {
      "post": {
        "operationId": "setSlippage",
        "summary": "Set default slippage tolerance",
        "description": "Operator-only endpoint to set the default slippage tolerance (0.1%–5%). Does not involve the chat LLM; the UI settings panel calls this directly.",
        "tags": ["Operator Settings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorSettingsRequest"
              },
              "example": {
                "vaultAddress": "0xYourVaultAddress",
                "chainId": 8453,
                "operatorAddress": "0xOperatorAddress",
                "authSignature": "0x...",
                "authTimestamp": 1741700000000,
                "slippage": "0.5%"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Slippage updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/settings/swap-shield": {
      "post": {
        "operationId": "setSwapShieldTolerance",
        "summary": "Set or reset Swap Shield tolerance",
        "description": "Operator-only endpoint to temporarily raise the Swap Shield tolerance (up to 50% for 10 minutes) or reset it to the default 5%. Does not involve the chat LLM.",
        "tags": ["Operator Settings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorSettingsRequest"
              },
              "example": {
                "vaultAddress": "0xYourVaultAddress",
                "chainId": 8453,
                "operatorAddress": "0xOperatorAddress",
                "authSignature": "0x...",
                "authTimestamp": 1741700000000,
                "tolerance": "30%"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Swap Shield tolerance updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/settings/nav-shield": {
      "post": {
        "operationId": "setNavShieldThreshold",
        "summary": "Set or reset NAV Shield threshold",
        "description": "Operator-only endpoint to set the NAV Shield maximum loss threshold (1%–100%) or reset it to the default 10%. Does not involve the chat LLM.",
        "tags": ["Operator Settings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorSettingsRequest"
              },
              "example": {
                "vaultAddress": "0xYourVaultAddress",
                "chainId": 8453,
                "operatorAddress": "0xOperatorAddress",
                "authSignature": "0x...",
                "authTimestamp": 1741700000000,
                "threshold": "15%"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NAV Shield threshold updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/settings/exec-mode": {
      "get": {
        "operationId": "getExecMode",
        "summary": "Get execution mode preference",
        "description": "Returns the operator's current auto-execute preference (autonomous or confirm). Shared between the web UI and Telegram.",
        "tags": ["Operator Settings"],
        "parameters": [
          { "name": "vaultAddress", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "chainId", "in": "query", "required": true, "schema": { "type": "integer" } },
          { "name": "operatorAddress", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "authSignature", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "authTimestamp", "in": "query", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Current execution mode preference",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecModeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "setExecMode",
        "summary": "Set execution mode preference",
        "description": "Sets whether the operator's trades execute immediately (autonomous) or require confirmation (confirm). Shared between the web UI and Telegram.",
        "tags": ["Operator Settings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecModeRequest"
              },
              "example": {
                "vaultAddress": "0xYourVaultAddress",
                "chainId": 8453,
                "operatorAddress": "0xOperatorAddress",
                "authSignature": "0x...",
                "authTimestamp": 1741700000000,
                "mode": "autonomous"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execution mode preference updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecModeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Payment",
        "description": "x402 payment header. Flow:\n1. Make the request without this header → receive `402 Payment Required` and, for protected routes, a `PAYMENT-REQUIRED` header describing accepted payment requirements\n2. Create the x402 payment payload for USDC on Base (chain 8453), typically via `@x402/core` and the CDP facilitator at `api.cdp.coinbase.com`\n3. Retry the request with this header set to the encoded payment payload\n\nSuccessful 2xx responses may include a `PAYMENT-RESPONSE` settlement receipt. See https://github.com/rigoblock/agentic-operator/blob/main/AGENTS.md for full integration."
      }
    },
    "schemas": {
...
      "OperatorSettingsRequest": {
        "type": "object",
        "required": ["vaultAddress", "chainId", "operatorAddress", "authSignature", "authTimestamp"],
        "properties": {
          "vaultAddress": {
            "type": "string",
            "description": "Vault contract address"
          },
          "chainId": {
            "type": "integer",
            "description": "Chain ID where the vault exists"
          },
          "operatorAddress": {
            "type": "string",
            "description": "Vault owner address"
          },
          "authSignature": {
            "type": "string",
            "description": "EIP-191 signature of the auth message"
          },
          "authTimestamp": {
            "type": "integer",
            "description": "Timestamp included in the signed auth message"
          },
          "slippage": {
            "type": "string",
            "description": "Slippage value (e.g. '0.5%', '50bps', or '0.5')"
          },
          "tolerance": {
            "type": "string",
            "description": "Swap Shield tolerance percentage (e.g. '30%')"
          },
          "threshold": {
            "type": "string",
            "description": "NAV Shield threshold percentage (e.g. '15%')"
          },
          "reset": {
            "type": "boolean",
            "description": "When true, reset the setting to its default value"
          }
        }
      },
      "ExecModeRequest": {
        "type": "object",
        "required": ["vaultAddress", "chainId", "operatorAddress", "authSignature", "authTimestamp", "mode"],
        "properties": {
          "vaultAddress": {
            "type": "string",
            "description": "Vault contract address"
          },
          "chainId": {
            "type": "integer",
            "description": "Chain ID where the vault exists"
          },
          "operatorAddress": {
            "type": "string",
            "description": "Vault owner address"
          },
          "authSignature": {
            "type": "string",
            "description": "EIP-191 signature of the auth message"
          },
          "authTimestamp": {
            "type": "integer",
            "description": "Timestamp included in the signed auth message"
          },
          "mode": {
            "type": "string",
            "enum": ["autonomous", "confirm"],
            "description": "Autonomous executes trades immediately; confirm requires an explicit approval"
          }
        }
      },
      "ExecModeResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "mode": {
            "type": "string",
            "enum": ["autonomous", "confirm"]
          }
        }
      },
      "SettingsResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
...
}